requirements.txt
python-dotenv
pandas
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-chroma
langchain-pinecone
pinecone-client
requests
fastapi